+++ src/app/(main)/_layout/Desktop/SideBar/Avatar.test.tsx
