--- src/app/(main)/_layout/type.ts
@@ -1,6 +1,0 @@
-import { ReactNode } from 'react';
-
-export interface LayoutProps {
-  children: ReactNode;
-  nav: ReactNode;
-}
