+++ src/app/(main)/layout.tsx
@@ -2,9 +2,8 @@ import ServerLayout from '@/components/server/ServerLayout';
 
 import Desktop from './_layout/Desktop';
 import Mobile from './_layout/Mobile';
-import { LayoutProps } from './_layout/type';
 
-const MainLayout = ServerLayout<LayoutProps>({ Desktop, Mobile });
+const MainLayout = ServerLayout({ Desktop, Mobile });
 
 MainLayout.displayName = 'MainLayout';
